requirements.txt
aws-cdk.aws-apigateway==1.159.0
aws-cdk.aws-appsync==1.159.0
aws-cdk.aws-backup==1.159.0
aws-cdk.aws-batch==1.159.0
aws-cdk.aws-cloudwatch==1.159.0
aws-cdk.aws-dynamodb==1.159.0
aws-cdk.aws-ec2==1.159.0
aws-cdk.aws-ecr==1.159.0
aws-cdk.aws-ecr-assets==1.159.0
aws-cdk.aws-ecs==1.159.0
aws-cdk.aws-elasticsearch==1.159.0
aws-cdk.aws-events==1.159.0
aws-cdk.aws-events-targets==1.159.0
aws-cdk.aws-iam==1.159.0
aws-cdk.aws-kinesis==1.159.0
aws-cdk.aws-kinesisfirehose==1.159.0
aws-cdk.aws-lambda==1.159.0
aws-cdk.aws-lambda-event-sources==1.159.0
aws-cdk.aws-logs==1.159.0
aws-cdk.aws-s3==1.159.0
aws-cdk.aws-s3-deployment==1.159.0
aws-cdk.aws-sqs==1.159.0
aws-cdk.aws-ssm==1.159.0
aws-cdk.aws-stepfunctions==1.159.0
aws-cdk.aws-stepfunctions-tasks==1.159.0
aws-cdk.cloudformation-include==1.159.0
aws-cdk.core==1.159.0
aws-cdk.aws-backup==1.159.0
aws-cdk.custom_resources==1.159.0

awscli
boto3
botocore
cdk-ecr-deployment==0.0.80
click
gql
jsii
jsonpickle==2.0.0
aiohttp
requests
py-cpuinfo
psutil